better kill test

+++ deploy/update.bash
@@ -43,8 +43,11 @@ echo "REMOTE DEPLOY: RESTARTING CELERY"
 celery multi stop w1 -A celery_server --logfile=w1.log --pidfile=w1.pid
 celery purge -f -A celery_server
 celery multi start w1 -A celery_server --logfile=w1.log --pidfile=w1.pid --loglevel=DEBUG
-kill $( cat beat.pid )
-rm beat.pid
+if [ -e beat.pid ]
+then
+  kill $( cat beat.pid );
+  rm beat.pid;
+fi
 celery -A celery_server beat --logfile=beat.log --pidfile=beat.pid --detach
 
 echo "REMOTE DEPLOY: STARTING SERVICE"
